型定義ディレクトリを追加
